Develop (#12)
* [update] update run

* [update] update nav

* [update] api url

* [feat] execute

* Squashed commit of the following:

commit c32d5cba1d0f605e0f3559f850ca12e5d3044c44
Author: Minh Nguyên <76799726+3langn@users.noreply.github.com>
Date:   Sun Jan 8 10:33:44 2023 +0700

    Develop (#9)

    * [update] update run

    * [update] update nav


commit 1f46a533ea106f86233843aa5b25b71f9ff91c89
Merge: 1c3136c a219c9d
Author: Minh Nguyên <76799726+3langn@users.noreply.github.com>
Date:   Fri Jan 6 15:26:19 2023 +0700

    Merge pull request #8 from 3langn/develop

    Develop

commit 1c3136cab24bf4d558f30c7e63bb8793770ba04b
Merge: 3a74a90 deba777
Author: Minh Nguyên <76799726+3langn@users.noreply.github.com>
Date:   Mon Jan 2 19:13:47 2023 +0700

    Merge pull request #7 from 3langn/develop

    Develop

commit 3a74a908e001c0079afc0e01163b20c442a908f2
Merge: f46143e 20c6084
Author: Minh Nguyên <76799726+3langn@users.noreply.github.com>
Date:   Fri Dec 30 22:03:58 2022 +0700

    Merge pull request #6 from 3langn/develop

    [fix] lession -> lesson

commit f46143e6c5b418ffadb74cbc8d4f9f164dfd5579
Merge: c4d0c9b c400a8d
Author: Minh Nguyên <76799726+3langn@users.noreply.github.com>
Date:   Fri Dec 30 21:41:52 2022 +0700

    Merge pull request #5 from 3langn/develop

    [feat] preview course

commit c4d0c9bba0bacfa27717809cb2bc44f871bd2030
Merge: 90589a4 cfafedc
Author: Minh Nguyên <76799726+3langn@users.noreply.github.com>
Date:   Wed Dec 28 14:19:10 2022 +0700

    Merge pull request #4 from 3langn/develop

    update

commit 90589a47eba06be6f696693b2c0ee84e2e8eaa2b
Merge: 570c7ac 909d508
Author: Minh Nguyên <76799726+3langn@users.noreply.github.com>
Date:   Wed Dec 28 14:13:07 2022 +0700

    Merge pull request #3 from 3langn/develop

    Develop

commit 570c7acc7215ce01bbb28e58a59a647a910c2ab7
Merge: 9e07ec7 497d13c
Author: Minh Nguyên <76799726+3langn@users.noreply.github.com>
Date:   Wed Dec 28 11:08:02 2022 +0700

    Merge pull request #2 from 3langn/develop

    Develop

* [update] refactor UI

* [update] add .vscode

* [update] order cat, lesson

* [fix]

+++ src/components/Lesson/CategoryMoreOptions.tsx
@@ -17,9 +17,8 @@ import { useAppDispatch, useAppSelector } from '../../app/hooks';
 import Button from '../../common/Button';
 import { setSnackBar } from '../../features/auth/appSlice';
 import {
-  addCategory,
-  addLesson,
   deleteCategoryById,
+  incrementActionCount,
   selectCourse,
   setCategoryTitle,
 } from '../../features/auth/LessonNavSlice';
@@ -31,6 +30,7 @@ interface CategoryMoreOptionsProps {
   editMode?: boolean;
   isHoverParent?: boolean;
   category: CategoryResponse;
+  index: number;
 }
 
 const CategoryMoreOptions: FC<CategoryMoreOptionsProps> = (props) => {
@@ -83,13 +83,18 @@ const CategoryMoreOptions: FC<CategoryMoreOptionsProps> = (props) => {
 
   const handleDelete = async () => {
     try {
+      if (course.category.length === 1) {
+        dispatch(setSnackBar({ message: 'Can not delete last category', type: 'error' }));
+        return;
+      }
       await CodeSmoothApi.deleteCategoryById(props.category.id);
       dispatch(deleteCategoryById(props.category.id));
       dispatch(setSnackBar({ message: 'Delete category success', type: 'success' }));
     } catch (error) {
       console.log(error);
+    } finally {
+      setAnchorEl(null);
     }
-    setAnchorEl(null);
   };
 
   const handleAddCategory = async () => {
@@ -101,14 +106,16 @@ const CategoryMoreOptions: FC<CategoryMoreOptionsProps> = (props) => {
         catid,
         course.id,
         CourseCategoryType.SECTION,
+        props.index + 1,
       );
-      dispatch(
-        addCategory({ title: inputNewCate, id: catid, currentCategoryId: props.category.id }),
-      );
+      // dispatch(
+      //   addCategory({ title: inputNewCate, id: catid, currentCategoryId: props.category.id }),
+      // );
+      dispatch(incrementActionCount());
       dispatch(setSnackBar({ message: 'Create category success', type: 'success' }));
       setAnchorEl(null);
     } catch (error: any) {
-      console.log(error);
+      dispatch(setSnackBar({ message: 'Create category failed', type: 'error' }));
     }
     setAnchorEl(null);
     setOpenDialogInputNewCategory(false);
@@ -117,19 +124,23 @@ const CategoryMoreOptions: FC<CategoryMoreOptionsProps> = (props) => {
   const handleAddLesson = async () => {
     try {
       const newLesson = generateLesson(props.category.id);
-      await CodeSmoothApi.saveLesson(newLesson);
-      dispatch(
-        addLesson({
-          category_id: props.category.id,
-          id: newLesson.id,
-          title: newLesson.title,
-        }),
-      );
-      dispatch(setSnackBar({ message: 'Create lesson success', type: 'success' }));
+      await CodeSmoothApi.addLesson({ ...newLesson, order: 0 });
+      // dispatch(
+      //   addLesson({
+      //     category_id: props.category.id,
+      //     id: newLesson.id,
+      //     title: newLesson.title,
+      //     index: 1,
+      //   }),
+      // );
+
+      dispatch(setSnackBar({ message: 'Create category success', type: 'success' }));
+      setAnchorEl(null);
     } catch (error: any) {
-      dispatch(setSnackBar({ message: error.message, type: 'error' }));
+      console.log(error);
     }
     setAnchorEl(null);
+    setOpenDialogInputNewCategory(false);
   };
 
   return (props.editMode && props.isHoverParent) || openMore || openDialogInputTitle ? (
